feat(Animation): animate :)

+++ src/framework/sketch.rs
@@ -3,7 +3,7 @@ use super::controls::Controls;
 pub struct SketchConfig {
     pub name: &'static str,
     pub display_name: &'static str,
-    pub fps: f64,
+    pub fps: f32,
     pub bpm: f32,
     pub w: i32,
     pub h: i32,
